requirements.txt
scikit-learn
joblib
flask
numpy
pillow
gunicorn
